// packages/fe-exam-pc/src/pages/element-surface/[type]/editTmp/AddModal/interface.d.ts
export type optionsType = {
    code: string
    name: string | undefined
    rate: string | undefined
    info: string | undefined
    type?: number
}
export type optionType = {
    name?: string | undefined
    rate?: string | undefined
    important?: string | undefined
    info?: string | undefined
}